python/requirements.txt
sentence-transformers==2.7.0
scikit-learn==1.5.1
numpy==1.24.3
pandas==2.0.3
hdbscan==0.8.33
umap-learn==0.5.5